Check the fit of a regression model used to estimate EVPPI or EVSI

Produces diagnostic plots and summaries of regression models used to estimate EVPPI or EVSI, mainly in order to check that the residuals have mean zero.

check_regression(
  x,
  pars = NULL,
  n = NULL,
  comparison = 1,
  outcome = "costs",
  plot = TRUE
)

Arguments

x

Output from evppi or evsi. The argument check=TRUE must have been used when calling evppi or evsi, to allow the regression model objects from gam or earth to be preserved. (This is not done by default, since these objects can be large.). attr(x, "models") cont

pars

Parameter (or parameter group) whose EVPPI calculation is to be checked. This should be in the pars component of the object returned by evppi. Only relevant if x is the result of an evppi calculation.

n

Sample size whose EVSI calculation is to be checked. This should be in the n component of the object returned by evsi. Only relevant if x is the result of an evsi calculation.

comparison

Only relevant if there are more than two treatments in the decision model. Different regression models are then used for the comparisons of different treatments with the baseline treatment. comparison is an integer identifying which of these models is checked.

outcome

"costs" or "effects". Only relevant if outputs was in cost-effectiveness format when calling evppi or evsi, hence different regressions are used for costs and effects. By default, outcome="costs" is used, so that the regression for costs is checked.

plot

If FALSE, only numerical statistics are returned, and a plot is not made.

Value

Where possible, an appropriate statistic is returned that allows the regression model to be compared with other regression models implemented using the same method

but with different assumptions. For method=="gam", this is Akaike's information criterion (AIC). For method=="earth", this is the generalised cross-validation statistic gcv. Currently not implemented for other methods.

Details

For VoI estimation, the key thing we are looking for is that the residuals have mean zero, hence that the mean of the model output is represented well by the regression function of the model input parameters. It should not matter if the variance of the residuals is non-constant, or non-normally distributed.

Models produced with method="gam" are checked using gam.check

Models produced method="earth" are checked using plot.earth

Models produced with method="gp" or method="inla" are checked using a histogram of the residuals, and plots of the residuals and response against the fitted values.

Examples

# TODO and refer in vignette