Chemotherapy health economic model

Calculate expected costs and effects for two treatments as a function of the input parameters. Each input parameter is a scalar.

chemo_costeff_fn(
  SE1,
  SE2,
  lambda.home.home,
  lambda.home.hosp,
  lambda.home.rec,
  lambda.hosp.hosp,
  lambda.hosp.rec,
  lambda.hosp.dead,
  c.home,
  c.hosp,
  c.dead,
  e.chemo,
  e.home,
  e.hosp
)

Arguments

SE1

TODO

SE2

TODO

lambda.home.home

Transition probability from

lambda.home.hosp

Transition probability from

lambda.home.rec

Transition probability from

lambda.hosp.hosp

Transition probability from

lambda.hosp.rec

Transition probability from

lambda.hosp.dead

Transition probability from

c.home

Cost TODO

c.hosp

Cost TODO

c.dead

Cost TODO

e.chemo

Effect TODO

e.home

Effect TODO

e.hosp

Effect TODO

Details

TODO refer to example of looping to generate PSA output matrix (currently in data-raw)