Chemotherapy cost-effectiveness model

An artificial health economic decision model with a typical Markov model structure, used for illustrating Value of Information methods. Functions are provided to generate model parameters and evaluate the model, and samples from probabilistic analysis of the model are provided as built-in datasets.

chemo_cea

chemo_nb

chemo_pars

chemo_pars_fn(n)

chemo_model_nb(
  p_side_effects_t1,
  p_side_effects_t2,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)

chemo_model_cea(
  p_side_effects_t1,
  p_side_effects_t2,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)

chemo_model_lor_nb(
  p_side_effects_t1,
  logor_side_effects,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)

chemo_model_lor_cea(
  p_side_effects_t1,
  logor_side_effects,
  p_home_home,
  p_home_hospital,
  p_home_recover,
  p_hospital_hospital,
  p_hospital_recover,
  p_hospital_dead,
  c_home_care,
  c_hospital,
  c_death,
  u_recovery,
  u_home_care,
  u_hospital
)

Format

Samples of 10000 from probabilistic analysis of this model are made available in the package, in the following data objects:

chemo_pars: Sample from the distributions of the parameters, as a data frame with names as documented above.

chemo_cea: List with components e (sampled effects), c (sampled costs), and k (a set of five equally-spaced willingess-to-pay values from 10000 to 50000 pounds). The effects and costs are data frames with two columns, one for each decision option.

chemo_nb: Data frame with two columns, giving the net monetary benefit for each decision option, at a willingness-to-pay of 20000 pounds.

Arguments

n

Number of samples to generate from the uncertainty distribution of the parameters in chemo_pars_fn.

p_side_effects_t1

Probability of side effects under treatment 1

p_side_effects_t2

Probability of side effects under treatment 2

p_home_home

Annual transition probability from home to home

p_home_hospital

Transition probability from home to hospital

p_home_recover

Transition probability from home to recovery

p_hospital_hospital

Transition probability from hospital to hospital

p_hospital_recover

Transition probability from hospital to recovery

p_hospital_dead

Transition probability from hospital to death

c_home_care

Cost of a yearly period under treatment at home

c_hospital

Cost of hospital treatment

c_death

Cost of death

u_recovery

Utility of a period in the recovery state

u_home_care

Utility of home care state

u_hospital

Utility of hospital state

logor_side_effects

Log odds ratio for side effects between treatment 1 and treatment 2

Value

Two alternative functions are provided to evaluate the decision model for given parameters.

chemo_model_nb returns a vector with elements giving the net monetary benefit for standard of care and novel treatment, respectively, at a willingness-to-pay of 20,000 pounds per QALY.

chemo_model_cea returns a matrix with:

two rows, the first for expected costs and the second for expected effects (QALYs) over the fifty year time horizon, and

two columns, the first for the "standard of care" decision option, and the second for the novel treatment.

chemo_model_lor_nb and chemo_model_lor_cea are the same model, but parameterised in terms of the probability of side effects for the standard of care p_side_effects_t1 and the log odds ratio of side effects between treatment groups logor_side_effects, rather than in terms of p_side_effects_t1 and p_side_effects_t2

chemo_pars_fn generates a sample from the uncertainty distribution of the parameters in the chemotherapy model . This returns a data frame with parameters matching the arguments of chemo_model_nb, and the following additional derived parameters:

logor_side_effects: log odds ratio of side effects for treatment 2 compared to 1

p_hospitalised_total: probability of hospitalisation over the 50 year time horizon

p_died: probability of death over the time horizon, given hospitalisation

lambda_home: conditional probability that a patient recovers given they are not hospitalised

lambda_hosp: conditional probability that a patient in hospital recovers given they do not die